Don't add empty elements to the list of configs.
Previusly, we added strings.Split() which when there is no "," was
returning the original string which was "".

This mean that we got an empty element in the set, and so the generated
annotation looked like ,test-ns/test-name with a leading comma for the
empty element.

This strips out empty elements, this has no technical change, it's
purely aesthetic.